requirements.txt
Pyrogram
TgCrypto
requests
facebook-sdk